done, not working nahui

+++ PointOfSaleSystem.API/Models/GiftCard.cs
@@ -5,6 +5,6 @@ namespace PointOfSaleSystem.API.Models
     public class GiftCard : BaseModel
     {
         public CurrencyEnum Currency { get; set; }
-        public float amount { get; set; } // percentage if tax, sum if 
+        public decimal amount { get; set; } // percentage if tax, sum if 
     }
 }
